+++ src/main/java/woowacourse/auth/domain/LoginMemberPrincipal.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.support;
+package woowacourse.auth.domain;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
