+++ src/main/java/woowacourse/config/ArgumentResolverConfig.java
@@ -4,27 +4,14 @@ import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import woowacourse.config.interceptor.LoginMemberArgumentResolver;
-import woowacourse.config.resolver.LoginMemberInterceptor;
+import woowacourse.config.resolver.LoginMemberArgumentResolver;
 
 @Configuration
 public class ArgumentResolverConfig implements WebMvcConfigurer {
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
         argumentResolvers.add(loginMemberArgumentResolver());
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(loginMemberInterceptor())
-                .addPathPatterns("/api/customer");
-    }
-
-    @Bean
-    public LoginMemberInterceptor loginMemberInterceptor() {
-        return new LoginMemberInterceptor();
     }
 
     @Bean
