+++ src/main/java/woowacourse/config/resolver/LoginMemberArgumentResolver.java
@@ -1,12 +1,13 @@
-package woowacourse.config.interceptor;
+package woowacourse.config.resolver;
 
+import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import woowacourse.auth.support.LoginMemberPrincipal;
+import woowacourse.auth.domain.LoginMemberPrincipal;
 import woowacourse.auth.ui.dto.LoginMember;
 
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
@@ -19,9 +20,11 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         final HttpServletRequest httpServletRequest = (HttpServletRequest) webRequest.getNativeRequest();
+        final Map<String, Object> payload = (Map<String, Object>) httpServletRequest.getAttribute("payload");
+        final long id = Long.parseLong(String.valueOf(payload.getOrDefault("id", 0)));
 
         try {
-            return LoginMember.from((long) httpServletRequest.getAttribute("id"));
+            return LoginMember.from(id);
         } catch (Exception e) {
             throw new IllegalArgumentException("로그인 아이디.... 리졸빙하다 오류...");
         }
