+++ src/main/java/woowacourse/shoppingcart/ui/CustomerController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import woowacourse.auth.support.LoginMemberPrincipal;
+import woowacourse.auth.domain.LoginMemberPrincipal;
 import woowacourse.auth.ui.dto.LoginMember;
 import woowacourse.shoppingcart.service.CustomerService;
 import woowacourse.shoppingcart.ui.dto.request.CustomerDeleteRequest;
